# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthcare Launchpad API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Job and internship listings
  - name: profiles
    description: Student profiles
  - name: employers
    description: Employer accounts
  - name: applications
    description: Job applications
  - name: dashboard
    description: Dashboard and analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List all job/internship listings
      parameters:
        - name: location
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: specialty
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a specific job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get the current user's student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertMyProfile
      tags: [profiles]
      summary: Create or update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileBody"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

  /profiles:
    get:
      operationId: getProfiles
      tags: [profiles]
      summary: Search/list student profiles (for employers)
      parameters:
        - name: specialty
          in: query
          schema:
            type: ["string", "null"]
        - name: location
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of student profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"

  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a specific student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /employers/me:
    get:
      operationId: getMyEmployer
      tags: [employers]
      summary: Get the current employer's account
      responses:
        "200":
          description: Employer account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertMyEmployer
      tags: [employers]
      summary: Create or update employer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertEmployerBody"
      responses:
        "200":
          description: Saved employer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employer"

  /applications:
    get:
      operationId: getApplications
      tags: [applications]
      summary: Get applications for the current user (student or employer view)
      parameters:
        - name: jobId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply to a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status (employer action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard stats (totals, counts by type/status)
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /jobs/featured:
    get:
      operationId: getFeaturedJobs
      tags: [jobs]
      summary: Get featured/highlighted job listings
      responses:
        "200":
          description: Featured jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        location:
          type: string
        type:
          type: string
          description: "internship | part-time | full-time | clinical"
        specialty:
          type: string
          description: "nursing | medical-assistant | emt | radiology | pharmacy | administration | other"
        description:
          type: string
        requirements:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        employerId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - company
        - location
        - type
        - specialty
        - description
        - isFeatured
        - createdAt

    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        type:
          type: string
        specialty:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        isFeatured:
          type: boolean
      required:
        - title
        - company
        - location
        - type
        - specialty
        - description
        - isFeatured

    UpdateJobBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        specialty:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        isFeatured:
          type: ["boolean", "null"]

    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        fullName:
          type: string
        email:
          type: string
        school:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        specialty:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkUserId
        - fullName
        - email
        - createdAt

    UpsertProfileBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        school:
          type: ["string", "null"]
        graduationYear:
          type: ["integer", "null"]
        specialty:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
      required:
        - fullName
        - email

    Employer:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        organizationName:
          type: string
        contactEmail:
          type: string
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          description: "hospital | clinic | urgent-care | nursing-home | other"
        createdAt:
          type: string
      required:
        - id
        - clerkUserId
        - organizationName
        - contactEmail
        - createdAt

    UpsertEmployerBody:
      type: object
      properties:
        organizationName:
          type: string
        contactEmail:
          type: string
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
      required:
        - organizationName
        - contactEmail

    Application:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        profileId:
          type: integer
        status:
          type: string
          description: "pending | reviewed | shortlisted | rejected | accepted"
        coverLetter:
          type: ["string", "null"]
        job:
          $ref: "#/components/schemas/Job"
        profile:
          $ref: "#/components/schemas/StudentProfile"
        createdAt:
          type: string
      required:
        - id
        - jobId
        - profileId
        - status
        - createdAt

    CreateApplicationBody:
      type: object
      properties:
        jobId:
          type: integer
        coverLetter:
          type: ["string", "null"]
      required:
        - jobId

    UpdateApplicationBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DashboardStats:
      type: object
      properties:
        totalJobs:
          type: integer
        totalStudents:
          type: integer
        totalApplications:
          type: integer
        totalEmployers:
          type: integer
        jobsByType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
            required:
              - type
              - count
        applicationsByStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required:
              - status
              - count
        jobsBySpecialty:
          type: array
          items:
            type: object
            properties:
              specialty:
                type: string
              count:
                type: integer
            required:
              - specialty
              - count
      required:
        - totalJobs
        - totalStudents
        - totalApplications
        - totalEmployers
        - jobsByType
        - applicationsByStatus
        - jobsBySpecialty

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          description: "new_job | new_application | application_updated | new_student"
        message:
          type: string
        createdAt:
          type: string
      required:
        - id
        - type
        - message
        - createdAt